+++ dart/geobase/CHANGELOG.md
@@ -12,6 +12,7 @@ NOTE: Version 1.1.0 currently under development (1.1.0-dev.0).
   * Support EWKB dimensionality flags for z and m.
   * CoordRefSys to support getting EPSG code as an integer.
   * Support encoding SRID to EWKB binary data.
+  * Utilities to decode a byte order, a WKB flavor and SRID from WKB bytes.
 
 ## 1.0.2